requirements.txt
# requirements.txt
flask
flask-cors
google-genai # New: For the Gemini API
gunicorn     # New: Production web server for Cloud Run